Added support for Kongsberg EM2040 multibeam in formats 58 & 59.
The function mb_get_info() now properly applies the lonflip value. This
in turn allows mbgrid to infer correct bounds in situations where the
lonflip is important (e.g. data across the +/-180 degree boundary when
lonflip is -1 or 1).

Changed MBnavadjust so that an "Inversion Smoothing" value replaces a 
no-longer-used "Inversion Precision" value. This value, accessible under
the Options dialog, sets the importance of smoothing in the adjusted
navigation model. The default is 4.00, and possible ranges are 0.10 to
10.0, where the smoothing penalty weight applied to the first and
second derivatives of the adjusted navigation perturbation is ten
raised to this power.


git-svn-id: svn://svn.ilab.ldeo.columbia.edu/repo/mb-system/trunk@1865 3bb7f1f9-8866-0410-989a-a36d4a1b0156